Refactored Teraz Gramy widget on the main page & moved into separate file
